requirements.txt
PyMuPDF>=1.23.0
Pillow>=10.0.0
requests>=2.31.0
numpy>=1.24.0
scikit-learn>=1.3.0
tqdm>=4.65.0
python-dotenv>=1.0.0
anthropic>=0.39.0
